+++ drivers/IMXRT1062/grblHAL_Teensy4/src/networking/WsStream.c
@@ -1,7 +1,7 @@
 //
 // WsStream.c - lwIP websocket stream implementation
 //
-// v1.2 / 2021-01-10 / Io Engineering / Terje
+// v1.3 / 2021-03-13 / Io Engineering / Terje
 //
 
 /*
@@ -256,6 +256,11 @@ void WsStreamRxCancel (void)
     streamSession.rxbuf.head = (streamSession.rxbuf.tail + 1) & (RX_BUFFER_SIZE - 1);
 }
 
+bool WsStreamSuspendInput (bool suspend)
+{
+    return stream_rx_suspend(&streamSession.rxbuf, suspend);
+}
+
 bool WsStreamRxInsert (char c)
 {
     // discard input if MPG has taken over...
@@ -265,7 +270,10 @@ bool WsStreamRxInsert (char c)
 
         if(bptr == streamSession.rxbuf.tail)                        // If buffer full
             streamSession.rxbuf.overflow = true;                    // flag overflow
-        else if(!hal.stream.enqueue_realtime_command(c)) {          // If not a real time command
+        else if(c == CMD_TOOL_ACK && !streamSession.rxbuf.backup) {
+            stream_rx_backup(&streamSession.rxbuf);
+            hal.stream.read = WsStreamGetC; // restore normal input
+        } if(!hal.stream.enqueue_realtime_command(c)) {             // If not a real time command
             streamSession.rxbuf.data[streamSession.rxbuf.head] = c; // add data to buffer
             streamSession.rxbuf.head = bptr;                        // and update pointer
         }