+++ drivers/IMXRT1062/grblHAL_Teensy4/src/uart.c
@@ -3,7 +3,7 @@
 
   Part of grblHAL
 
-  Some parts of this code is Copyright (c) 2020 Terje Io
+  Some parts of this code is Copyright (c) 2020-2021 Terje Io
 
   Some parts are derived from HardwareSerial.cpp in the Teensyduino Core Library
 
@@ -141,7 +141,7 @@ void transmitterEnable(uint8_t pin)
 
 static uint16_t tx_fifo_size;
 static stream_tx_buffer_t txbuffer = {0};
-static stream_rx_buffer_t rxbuffer = {0}, rxbackup;
+static stream_rx_buffer_t rxbuffer = {0};
 
 void serialInit (uint32_t baud_rate)
 {
@@ -339,20 +339,9 @@ void serialWrite(const char *s, uint16_t length)
         serialPutC(*ptr++);
 }
 
-// "dummy" version of serialGetC
-static int16_t serialGetNull (void)
-{
-    return -1;
-}
-
 bool serialSuspendInput (bool suspend)
 {
-    if(suspend)
-        hal.stream.read = serialGetNull;
-    else if(rxbuffer.backup)
-        memcpy(&rxbuffer, &rxbackup, sizeof(stream_rx_buffer_t));
-
-    return rxbuffer.tail != rxbuffer.head;
+    return stream_rx_suspend(&rxbuffer, suspend);
 }
 
 uint16_t serialTxCount(void) {
@@ -413,9 +402,7 @@ static void uart_interrupt_handler (void)
                 rxbuffer.head = bptr;                       // and update pointer
 #else
                 if(data == CMD_TOOL_ACK && !rxbuffer.backup) {
-                    memcpy(&rxbackup, &rxbuffer, sizeof(stream_rx_buffer_t));
-                    rxbuffer.backup = true;
-                    rxbuffer.tail = rxbuffer.head;
+                    stream_rx_backup(&rxbuffer);
                     hal.stream.read = serialGetC; // restore normal input
                 } else if(!hal.stream.enqueue_realtime_command((char)data)) {
                     rxbuffer.data[rxbuffer.head] = (char)data;  // Add data to buffer