+++ drivers/RP2040/generic_map.h
@@ -114,10 +114,3 @@
 #define KEYPAD_STROBE_PIN           15
 #define KEYPAD_STROBE_BIT           (1<<KEYPAD_STROBE_PIN)
 #endif
-
-#if SDCARD_ENABLE
-#define SD_CS_PORT  GPIOA
-#define SD_CS_PIN   3
-#define SD_CS_BIT   (1<<SD_CS_PIN)
-#define SPI_PORT    1 // GPIOA, SCK_PIN = 5, MISO_PIN = 6, MOSI_PIN = 7
-#endif